backend/requirements.txt
easyocr==1.6.2
matplotlib==3.6.2
opencv-python-headless==4.5.4.60
flask
numpy<2
pillow